backend/requirements.txt
requests==2.26.0
Django==3.2.16
djangorestframework==3.12.4
django-filter==2.4.0
django-rest-swagger==2.2.0
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
asgiref==3.5.2  
python-dotenv==0.19.0
djoser==2.1.0
Pillow==8.3.1
isort==5.11.4
drf-extra-fields==3.2.1
drf-base64==2.0